requirements.txt
cbor2==5.6.5
luma.core==2.4.2
luma.oled==3.14.0
pillow==11.1.0
pyftdi==0.56.0
pyserial==3.5
pyusb==1.3.1
RPi.GPIO==0.7.1
smbus2==0.5.0
spidev==3.6
